Set minimum version to 6.3
Apparently on windows the version number is the original version
of windows.

If a user upgrades from windows 8.1 to windows 10 their computer
still thinks that the windows version number is 6.3

I have no idea why its 6.3 instead of 10; version numbers of
the OS on windows make zero sense to me.