cli: don't exit() on success
Running exit() from another thread on Windows will shutdown the entire
process. So only exit the process early (which will cancel the build of
ssc.exe) if there was an error running the build script.

This is not a cross-platform compatible solution. Running exit() from a
different thread on Linux will only exit the thread. Futher work needs
to be done to make this work everywhere.

PR-URL: https://github.com/socketsupply/socket/pull/55
Reviewed-by: Joseph Werle <joseph.werle@gmail.com>
Reviewed-by: heapwolf <paolo@socketsupply.co>