src: remove commented code + add comments
Remove commented code and replace with TODO tags. Most the commented
code had to do with dark mode implementation.

Add comment about which header contains the definitions for message in
the call to WndProc().

Use CreateSolidBrush() directly instead of trying to set it as a
variable. Making it more difficult to access.

PR-URL: https://github.com/socketsupply/socket/pull/55
Reviewed-by: Joseph Werle <joseph.werle@gmail.com>
Reviewed-by: heapwolf <paolo@socketsupply.co>