window: use unique app ptr in lambda
When shutdownHandler called from the windowManager instance (assigned as
onExit) the member variable "app" is used instead of the stack allocated
App instance.

Trying to access the member variable app after running
windowManager.destroy() causes a segfault since the reference on the
WindowManager instance is no longer valid.

To get around this assign the App instance to a static pointer variable
and use that instead.

Make the App and WindowManager instances static to simplify
initialization order and remove the possibility for a race condition.

PR-URL: https://github.com/socketsupply/socket/pull/66
Reviewed-by: Sergey Rubanov <chi187@gmail.com>