window: implement bridge routing
Handle bridge routing in the case of an XHR request. To do this need to
detect the preflight OPTION request. Then handle the GET or POST request
afterwards.

Need to finish the response in a call to dispatch() because completion
must be done in the same thread that made the request. This makes the
assumption that all requests are done from the main thread.

TODO: If an XHR request is made from not the main thread then this will
break. The call to put_Response() will fail with UI_E_WRONG_THREAD
(0x802A000C).

Attempt a bridge route call when a web message is received. If it fails
then forward the message to onMessage().

PR-URL: https://github.com/socketsupply/socket/pull/71
Reviewed-by: Joseph Werle <joseph.werle@gmail.com>