windows: support custom ipc: schemes
Use WebView2's custom scheme registration to accept ipc: in XHR
requests. The allowed origins were set very broad since these requests
should only be made internally from the JS API.

Custom schemes have a bug. When trying to receive the POST body of a
request, a random pointer is instead returned. To get around this we use
the shared buffer API to temporarily store the POST body, and retrieve
it later when the POST request is made.

The current solution for storing the POST body is a race condition that
could lead to a memory leak. If the shared buffer is placed in the
message map, and the window is closed, there is no mechanism that will
clean out the stored maps.

Proper cleanup probably requires accessing the map and removing the
shared buffers specific to that window before it finishes closing. This
is because accessing the shared buffer, which is associated to a
specific window, will be invalid after the window has closed.

But this isn't a problem at the moment since Windows doesn't currently
support more than the main window. So at the moment it's impossible to
leak memory.

Ref: https://github.com/MicrosoftEdge/WebView2Feedback/issues/2294
PR-URL: https://github.com/socketsupply/socket/pull/157
Reviewed-by: Joseph Werle <joseph.werle@gmail.com>