test: fix dirname path in Windows
Converting the dirname from file:/// would leave in /C:/ on Windows. Add
a replace to remove any drive letter from the remaining path and give a
unix-like path.

Note: This script still can't be reliably run on Windows. It will cause
the test to start in the background and stall. But it is still necessary
to run in order to place the fixtures in the correct location.

PR-URL: https://github.com/socketsupply/socket/pull/160
Reviewed-by: Joseph Werle <joseph.werle@gmail.com>