fs: remove iov buf vector
Only the first uv_buf_t was ever used in fs. Remove it to simplify
memory management.

Change setBuffer() to accept uint32_t since uv_buf_t.len in Windows uses
a ULONG.

Add comment in ~RequestContext() as a reminder whether we should be
free'ing that memory or not. I'm still not clear on whether the
RequestContext owns that memory, or if ownership has been given to the
user.