fix(npm index.js): convert file url to actual path
Certainly on Windows (but maybe on other platforms too), file URLs can contain spaces which will be encoded as %20. When then treating these strings as paths, the OS will claim it doesn't recognize these paths because it won't URL-decode them automatically. Therefore, the `import.meta.url` needs to be converted to an actual filePath before all the other path processing can be done on it.